Fix bug when upgrading opsman appliance
From time to time pivotal depprecates minor versions from pivotal
network. When ops_manager CLI tries to download this versions that it
was using on the pre upgrade ops manager appliance it used to fail.
Now it tries to download that version, if it succeeds it will upload the
exact same version to the new ops manager, if not it will try to match
the major version while looking for a grater minor one.

Signed-off-by: Lukasz Rabczak <lukasz.rabczak@gmail.com>